0.4.0: allow ansi-colored place names
The map will now be drawn by checking the printable (ascii) characters
of the place name, allowing for ansi characters to be present in their
name (and the map still show up nicely).

This includes a dependency of strip-ansi, which is marked as <7 on
node's package.json but not on debian's control file, as the version
limitation is there just to overcome what I consider a design bug of
npm. More info about that at
https://github.com/marado/TalkerNode/commit/ca64fcd0704f024b5d55290d57682d5888593dd0